added admin guard on admin specific APIs

+++ src/category/category.controller.ts
@@ -1,5 +1,16 @@
-import { Body, Controller, Delete, Get, HttpCode, HttpStatus, Param, Post } from '@nestjs/common';
-import { Public } from '../common/decorators';
+import {
+    Body,
+    Controller,
+    Delete,
+    Get,
+    HttpCode,
+    HttpStatus,
+    Param,
+    Post,
+    // eslint-disable-next-line prettier/prettier
+    UnauthorizedException
+} from '@nestjs/common';
+import { GetCurrentUser, Public } from '../common/decorators';
 import { CategoryService } from './category.service';
 import { CategoryDto, CategoryUpdateDto, SubCategoryUpdateDto } from './dto';
 import { Category } from './schema';
@@ -10,7 +21,14 @@ export class CategoryController {
 
     @Post()
     @HttpCode(HttpStatus.CREATED)
-    create(@Body() dto: CategoryDto): Promise<Category> {
+    create(
+        @GetCurrentUser('isAdmin') isAdmin: boolean,
+        @Body() dto: CategoryDto
+    ): Promise<Category> {
+        if (!isAdmin) {
+            throw new UnauthorizedException('you are not the admin');
+        }
+
         return this.categoryService.create(dto);
     }
 
@@ -30,32 +48,57 @@ export class CategoryController {
 
     @Post('/:id')
     @HttpCode(HttpStatus.OK)
-    update(@Param('id') id: string | number, @Body() dto: CategoryUpdateDto): Promise<Category> {
+    update(
+        @GetCurrentUser('isAdmin') isAdmin: boolean,
+        @Param('id') id: string | number,
+        @Body() dto: CategoryUpdateDto
+    ): Promise<Category> {
+        if (!isAdmin) {
+            throw new UnauthorizedException('you are not the admin');
+        }
+
         return this.categoryService.update(id, dto);
     }
 
     @Delete('/:id')
     @HttpCode(HttpStatus.OK)
-    delete(@Param('id') id: string | number): Promise<Category> {
+    delete(
+        @GetCurrentUser('isAdmin') isAdmin: boolean,
+        @Param('id') id: string | number
+    ): Promise<Category> {
+        if (!isAdmin) {
+            throw new UnauthorizedException('you are not the admin');
+        }
+
         return this.categoryService.delete(id);
     }
 
     @Post('/:categoryId/:subCategoryId')
     @HttpCode(HttpStatus.OK)
     updateSubCategory(
+        @GetCurrentUser('isAdmin') isAdmin: boolean,
         @Param('categoryId') categoryId: string | number,
         @Param('subCategoryId') subCategoryId: string | number,
         @Body() dto: SubCategoryUpdateDto
     ): Promise<Category> {
+        if (!isAdmin) {
+            throw new UnauthorizedException('you are not the admin');
+        }
+
         return this.categoryService.updateSubCategory(categoryId, subCategoryId, dto);
     }
 
     @Delete('/:categoryId/:subCategoryId')
     @HttpCode(HttpStatus.OK)
     deleteSubCategory(
+        @GetCurrentUser('isAdmin') isAdmin: boolean,
         @Param('categoryId') categoryId: string | number,
         @Param('subCategoryId') subCategoryId: string | number
     ): Promise<Category> {
+        if (!isAdmin) {
+            throw new UnauthorizedException('you are not the admin');
+        }
+
         return this.categoryService.deleteSubCategory(categoryId, subCategoryId);
     }
 }
